refactor: rename to lexical editor action to match the component name
